+++ ghost/core/test/e2e-browser/admin/site-settings.spec.js
@@ -1,6 +1,6 @@
 const {expect} = require('@playwright/test');
 const test = require('../fixtures/ghost-test');
-const {createPostDraft, createTier, disconnectStripe, generateStripeIntegrationToken, setupStripe} = require('../utils');
+const {createPostDraft, createTier, disconnectStripe, generateStripeIntegrationToken, setupStripe, getStripeAccountId} = require('../utils');
 
 const changeSubscriptionAccess = async (page, access) => {
     await page.locator('[data-test-nav="settings"]').click();
@@ -126,7 +126,8 @@ test.describe('Site Settings', () => {
             await changeSubscriptionAccess(page, 'all');
 
             await page.goto('/ghost');
-            const stripeToken = await generateStripeIntegrationToken();
+            const stripeAccountId = await getStripeAccountId();
+            const stripeToken = await generateStripeIntegrationToken(stripeAccountId);
             await setupStripe(page, stripeToken);
         });
     });
